requirements.txt
streamlit>=1.28.0
tensorflow>=2.10.0
Pillow
numpy
opencv-python
matplotlib
plotly
scikit-learn
geopy
gdown
